Added DataValidation

+++ Api/Controllers/CommentController.cs
@@ -21,6 +21,11 @@ namespace Api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var comments = await this._commentRepository.GetAllAsync();
 
             var commentsDto = comments.Select(comm => comm.ToCommentDto());
@@ -31,6 +36,11 @@ namespace Api.Controllers
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var comment = await this._commentRepository.GetByIdAsync(id);
 
             return comment is null
@@ -41,6 +51,11 @@ namespace Api.Controllers
         [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto createCommentDto )
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             if (!await this._stockRepository.StockExists(stockId))
             {
                 return BadRequest("Stock doesn't exist");
@@ -56,6 +71,11 @@ namespace Api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateCommentRequestDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var comment = await this._commentRepository.UpdateAsync(id, updateCommentRequestDto.ToCommentFromUpdateDto());
 
             return comment is null
@@ -67,6 +87,11 @@ namespace Api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var commentModel = await this._commentRepository.DeleteAsync(id);
 
             if (commentModel is null)
